LayoutTests/inspector/protocol/console-agent.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/protocol-test.js"></script>
<script>

function test()
{
    var testSuite = [
        ["ConsoleAgent", "setConsoleMessagesEnabled", false],
        ["ConsoleAgent", "setConsoleMessagesEnabled", true],
        ["RuntimeAgent", "evaluate", "console.info('test'); console.info('test')", "test-group", false],
        ["ConsoleAgent", "clearConsoleMessages"],
        ["ConsoleAgent", "setMonitoringXHREnabled", true],
        ["ConsoleAgent", "setMonitoringXHREnabled", false],
        ["ConsoleAgent", "addInspectedNode", 1]
    ];

    InspectorTest.runProtocolTestSuite("ConsoleAgent", testSuite, ['id', 'requestId', 'injectedScriptId', 'stackTrace']);
}

</script>
</head>

<body onload="runTest()">
<p>
Protocol stability test. It is dumping request/response pairs of ConsoleAgent functions.
</p>

</body>
</html>
