LayoutTests/inspector/protocol/console-agent-expected.txt

CONSOLE MESSAGE: line 1: test
CONSOLE MESSAGE: line 1: test
Protocol stability test. It is dumping request/response pairs of ConsoleAgent functions.

-----------------------------------------------------------
ConsoleAgent.setConsoleMessagesEnabled(false)

request:
{
    id : <number>
    domain : "Console"
    command : "setConsoleMessagesEnabled"
    arguments : {
        enabled : false
    }
}

response:
{
    requestId : <number>
    body : {
        newState : false
    }
}

-----------------------------------------------------------
ConsoleAgent.setConsoleMessagesEnabled(true)

request:
{
    id : <number>
    domain : "Console"
    command : "setConsoleMessagesEnabled"
    arguments : {
        enabled : true
    }
}

response:
{
    requestId : <number>
    body : {
        newState : true
    }
}

-----------------------------------------------------------
RuntimeAgent.evaluate("console.info('test'); console.info('test')","test-group",false)

request:
{
    id : <number>
    domain : "Runtime"
    command : "evaluate"
    arguments : {
        expression : "console.info('test'); console.info('test')"
        objectGroup : "test-group"
        includeCommandLineAPI : false
    }
}

event ConsoleAgent.addConsoleMessage
{
    type : "event"
    domain : "Console"
    event : "addConsoleMessage"
    body : {
        messageObj : {
            source : 3
            type : 0
            level : 1
            line : 0
            url : ""
            repeatCount : 1
            message : "test"
            parameters : {
                0 : {
                    type : "string"
                    description : "test"
                }
            }
            stackTrace : <object>
        }
    }
}

event ConsoleAgent.updateConsoleMessageRepeatCount
{
    type : "event"
    domain : "Console"
    event : "updateConsoleMessageRepeatCount"
    body : {
        count : 2
    }
}

response:
{
    requestId : <number>
    body : {
        result : {
            type : "undefined"
            description : "undefined"
        }
    }
}

-----------------------------------------------------------
ConsoleAgent.clearConsoleMessages()

request:
{
    id : <number>
    domain : "Console"
    command : "clearConsoleMessages"
    arguments : {
    }
}

event ConsoleAgent.consoleMessagesCleared
{
    type : "event"
    domain : "Console"
    event : "consoleMessagesCleared"
    body : {
    }
}

response:
{
    requestId : <number>
}

-----------------------------------------------------------
ConsoleAgent.setMonitoringXHREnabled(false)

request:
{
    id : <number>
    domain : "Console"
    command : "setMonitoringXHREnabled"
    arguments : {
        enabled : false
    }
}

response:
{
    requestId : <number>
}

-----------------------------------------------------------
ConsoleAgent.setMonitoringXHREnabled(true)

request:
{
    id : <number>
    domain : "Console"
    command : "setMonitoringXHREnabled"
    arguments : {
        enabled : true
    }
}

response:
{
    requestId : <number>
}

-----------------------------------------------------------
ConsoleAgent.addInspectedNode(<number>)

request:
{
    id : <number>
    domain : "Console"
    command : "addInspectedNode"
    arguments : {
        nodeId : 1
    }
}

response:
{
    requestId : <number>
}

===========================================================
Coverage for ConsoleAgent
{
    setConsoleMessagesEnabled : "checked"
    clearConsoleMessages : "checked"
    setMonitoringXHREnabled : "checked"
    addInspectedNode : "checked"
    addConsoleMessage : "checked"
    updateConsoleMessageExpiredCount : "not checked"
    updateConsoleMessageRepeatCount : "checked"
    consoleMessagesCleared : "checked"
}
